pyproject.toml
[project]
name = "mcp-codebase-analyser"
version = "0.1.0"
description = "MCP server for intelligent codebase analysis and section recommendations"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.3.0",
    "gitpython>=3.1.40",
    "chromadb>=0.4.22",
    "openai>=1.12.0",
    "tiktoken>=0.5.2",
    "tree-sitter>=0.21.0",
    "tree-sitter-javascript>=0.21.0",
    "tree-sitter-typescript>=0.21.0",
    "tree-sitter-python>=0.21.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.2.1"
]
